# enh: optimizing the reading of eigenstates from WFSX file