--- a/sisl/io/siesta/sile.py
+++ b/sisl/io/siesta/sile.py
@@ -25,6 +25,8 @@ class SileCDFSiesta(SileCDF):
 
     # all netcdf output should not be masked
     def _setup(self, *args, **kwargs):
+        super()._setup(*args, **kwargs)
+
         # all NetCDF routines actually returns masked arrays
         # this is to prevent Siesta CDF files from doing this.
         if hasattr(self, "fh"):
@@ -36,18 +38,19 @@ class SileBinSiesta(SileBin):
 
     def _setup(self, *args, **kwargs):
         """We set up everything to handle the fortran I/O unit"""
+        super()._setup(*args, **kwargs)
         self._iu = -1
 
-    def _bin_check(self, method, message):
+    def _fortran_check(self, method, message):
         ierr = _siesta.io_m.iostat_query()
         if ierr != 0:
             raise SileError(f'{self!s}.{method} {message} (ierr={ierr})')
 
-    def _is_fortran_unit_open(self):
+    def _fortran_is_open(self):
         return self._iu != -1
 
-    def _open_fortran_unit(self, mode, rewind=False):
-        if self._is_fortran_unit_open() and mode == self._mode:
+    def _fortran_open(self, mode, rewind=False):
+        if self._fortran_is_open() and mode == self._mode:
             if rewind:
                 _siesta.io_m.rewind_file(self._iu)
             else:
@@ -60,10 +63,10 @@ class SileBinSiesta(SileBin):
                 self._iu = _siesta.io_m.open_file_write(self.file)
             else:
                 raise SileError(f"Mode '{mode}' is not an accepted mode to open a fortran file unit. Use only 'r' or 'w'")
-        self._bin_check('_open_fortran_unit', 'could not open for {}.'.format({'r': 'reading', 'w': 'writing'}[mode]))
-        
-    def _close_fortran_unit(self):
-        if not self._is_fortran_unit_open():
+        self._fortran_check('_fortran_open', 'could not open for {}.'.format({'r': 'reading', 'w': 'writing'}[mode]))
+
+    def _fortran_close(self):
+        if not self._fortran_is_open():
             return
         # Close it
         _siesta.io_m.close_file(self._iu)