… sisl/io/siesta/binaries.py
… sisl/viz/plots/bands.py
… sisl/viz/plots/grid.py
… sisl/viz/plots/pdos.py