Addresses https://github.com/zerothi/sisl/issues/147 by making python keep track of the fortran file unit.

Now if one wants to iterate over all eigenstates the whole file is only read once, in contrast to the previous implementation, which read the file from the beginning at each iteration. This quick code snippet shows the difference in performance for iterating over all eigenstates for the file `spin_unpolarized.bands.WFSX`, which contains 95 k points:

```python
import sisl

wfsx = sisl.get_sile("spin_unpolarized.bands.WFSX")
geom = sisl.geom.diamond()

def iterate(which):
    it = {"new": wfsx.yield_eigenstate, "old": wfsx.old_yield_eigenstate}[which]
    
    for state in it(parent=geom):
        pass
    
%timeit iterate("new")
%timeit iterate("old")
```
```
1.87 ms ± 90.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
35.7 ms ± 97.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

I have some doubts:
- I think the spin and k indices are not written on the WFSX file as the old implementation assumes (`spin-loop[k-loop]`). I believe it's the opposite (`k-loop[spin-loop]`). You can test this by trying to read a polarized WFSX file, which will fail on the second eigenstate. I attach one so that you can test it. If this is true, I need to make some changes to the old implementation.
- You already mentioned in https://github.com/zerothi/sisl/issues/147 that the reading is unstable. I think I stumble upon this when reading the polarized file attached here. The binary check raises an error at `ik = 68`. Why is this? :sweat_smile: 

Files (too big to attach): https://drive.google.com/drive/folders/1lgqckfkpq0mZo-3ucBxtXvz5fMKlXCQv?usp=sharing