@zerothi 2021-10-20
Thanks for your comments. Indeed this shows some inconsistencies in the entire SiestaBin suite!!! :)

I agree this might be fitting in the binary sile for siesta, for coherence! Thanks for noticing.

Regarding the `read_eigenstate`, this is probably my mistake!!! ;)
The idea should be that all methods have a similar interface, so it should definitely be `k` and not `ik`. So this needs changing. I just recall I did this *quickly* because somebody wanted to use the wavefunctions. So I think it is ok to break API to make it stable.