@pfebrer 2021-10-20
> Regarding the read_eigenstate, this is probably my mistake!!! ;)

No, no. It wasn't there. I added it and then removed it. But couldn't it be useful in some cases to be able to get an eigenstate by its index? It would use the old subroutines and therefore avoid going back and forth between fortran and python.