@pfebrer 2021-10-21
I added the possibility to read the information (not values) of all k points at once. It is useful for debugging but could be also useful for other things (you can get a `BandStructure` / `BrillouinZone` object easily for example, or find the best match to your desired k value).

Since all information is read in one call this is significantly faster than iterating over the eigenstates:

```python
import sisl
import numpy as np

wfsx = sisl.get_sile("spin_unpolarized.bands.WFSX")
geom = sisl.geom.diamond()
def all_fortran():
    ks = wfsx.read_file_info(parent=geom)

def python_loop():
    for state in wfsx.yield_eigenstate(parent=geom):
        state.info
        
%timeit all_fortran()
%timeit python_loop()
```
```
335 µs ± 9.52 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
1.99 ms ± 112 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
Here is an example of what is returned by `read_file_info`:
```
{'k': array([[0.     , 0.     , 0.     ],
        [0.     , 0.0714 , 0.0714 ],
        [0.     , 0.1428 , 0.1428 ],
        [0.     , 0.2142 , 0.2142 ],
        [0.     , 0.2856 , 0.2856 ],
        [0.     , 0.357  , 0.357  ],
        [0.     , 0.4284 , 0.4284 ],
        [0.     , 0.4998 , 0.4998 ],
        [0.     , 0.5712 , 0.5712 ],
        [0.     , 0.6426 , 0.6426 ],
        [0.     , 0.714  , 0.714  ],
        [0.     , 0.7854 , 0.7854 ],
        [0.     , 0.8568 , 0.8568 ],
        [0.     , 0.9282 , 0.9282 ],
        [0.     , 0.9996 , 0.9996 ],
        [0.     , 1.071  , 1.071  ],
        [0.     , 1.1424 , 1.1424 ],
        [0.     , 1.2138 , 1.2138 ],
        [0.     , 1.2852 , 1.2852 ],
        [0.     , 1.3566 , 1.3566 ],
        [0.     , 1.428  , 1.428  ],
        [0.     , 1.4994 , 1.4994 ],
        [0.     , 1.5708 , 1.5708 ],
        [0.     , 1.6422 , 1.6422 ],
        [0.     , 1.7136 , 1.7136 ],
        [0.     , 1.785  , 1.785  ],
        [0.08925, 1.785  , 1.87425],
        [0.1785 , 1.785  , 1.9635 ],
        [0.26775, 1.785  , 2.05275],
        [0.357  , 1.785  , 2.142  ],
        [0.44625, 1.785  , 2.23125],
        [0.5355 , 1.785  , 2.3205 ],
        [0.62475, 1.785  , 2.40975],
        [0.714  , 1.785  , 2.499  ],
        [0.80325, 1.785  , 2.58825],
        [0.8925 , 1.785  , 2.6775 ],
        [0.952  , 1.785  , 2.618  ],
        [1.0115 , 1.785  , 2.5585 ],
        [1.071  , 1.785  , 2.499  ],
        [1.1305 , 1.785  , 2.4395 ],
        [1.19   , 1.785  , 2.38   ],
        [1.2495 , 1.785  , 2.3205 ],
        [1.309  , 1.785  , 2.261  ],
        [1.3685 , 1.785  , 2.2015 ],
        [1.428  , 1.785  , 2.142  ],
        [1.4875 , 1.785  , 2.0825 ],
        [1.547  , 1.785  , 2.023  ],
        [1.6065 , 1.785  , 1.9635 ],
        [1.666  , 1.785  , 1.904  ],
        [1.7255 , 1.785  , 1.8445 ],
        [1.785  , 1.785  , 1.785  ],
        [1.69575, 1.69575, 1.69575],
        [1.6065 , 1.6065 , 1.6065 ],
        [1.51725, 1.51725, 1.51725],
        [1.428  , 1.428  , 1.428  ],
        [1.33875, 1.33875, 1.33875],
        [1.2495 , 1.2495 , 1.2495 ],
        [1.16025, 1.16025, 1.16025],
        [1.071  , 1.071  , 1.071  ],
        [0.98175, 0.98175, 0.98175],
        [0.8925 , 0.8925 , 0.8925 ],
        [0.80325, 0.80325, 0.80325],
        [0.714  , 0.714  , 0.714  ],
        [0.62475, 0.62475, 0.62475],
        [0.5355 , 0.5355 , 0.5355 ],
        [0.44625, 0.44625, 0.44625],
        [0.357  , 0.357  , 0.357  ],
        [0.26775, 0.26775, 0.26775],
        [0.1785 , 0.1785 , 0.1785 ],
        [0.08925, 0.08925, 0.08925],
        [0.     , 0.     , 0.     ],
        [0.1071 , 0.1071 , 0.1071 ],
        [0.2142 , 0.2142 , 0.2142 ],
        [0.3213 , 0.3213 , 0.3213 ],
        [0.4284 , 0.4284 , 0.4284 ],
        [0.5355 , 0.5355 , 0.5355 ],
        [0.6426 , 0.6426 , 0.6426 ],
        [0.7497 , 0.7497 , 0.7497 ],
        [0.8568 , 0.8568 , 0.8568 ],
        [0.9639 , 0.9639 , 0.9639 ],
        [1.071  , 1.071  , 1.071  ],
        [1.1781 , 1.1781 , 1.1781 ],
        [1.2852 , 1.2852 , 1.2852 ],
        [1.3923 , 1.3923 , 1.3923 ],
        [1.4994 , 1.4994 , 1.4994 ],
        [1.6065 , 1.6065 , 1.6065 ],
        [1.7136 , 1.7136 , 1.7136 ],
        [1.8207 , 1.8207 , 1.8207 ],
        [1.9278 , 1.9278 , 1.9278 ],
        [2.0349 , 2.0349 , 2.0349 ],
        [2.142  , 2.142  , 2.142  ],
        [2.2491 , 2.2491 , 2.2491 ],
        [2.3562 , 2.3562 , 2.3562 ],
        [2.4633 , 2.4633 , 2.4633 ],
        [2.5704 , 2.5704 , 2.5704 ],
        [2.6775 , 2.6775 , 2.6775 ]]),
 'weight': array([1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
        1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.]),
 'nwf': array([[8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
         8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
         8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
         8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8,
         8, 8, 8, 8, 8, 8, 8, 8]], dtype=int32)}
```

On the 505 MB file that I attached, reading the info takes 140 ms. So I'd say this is very usable in practice even for huge WFSX files.