@pfebrer 2021-10-21
Ok, all this is because I want to allow using WFSX files for plotting wavefunctions, (fat)bands and PDOS.

For the wavefunctions and (fat)bands I will just add entry points in the respective plots. For the PDOS, do you think it would be useful to provide a method outside the plotting suite as well? E.g. : `wfsxSileSiesta.read_PDOS()`. 

Or maybe in a more abstract way provide apply dispatchers that would work similarly as the `BrillouinZone` applies? (I like this idea)