@pfebrer 2021-11-08
Can I generate a `BandStructure` object from a list of k points?