@pfebrer 2021-11-09
Also, one question for PDOS. If I want to generate the PDOS from the `WFSX` file, I need to read the overlap matrix as well, right?