@zerothi 2021-11-09
> Can I generate a `BandStructure` object from a list of k points? I want to be able to plot bands and fatbands just from the `WFSX` file, but I am missing some points sweat_smile

Yes, I can't see why not. Except that the `bandstructure` object requires a parent with cell information.

> Also, one question for PDOS. If I want to generate the PDOS from the `WFSX` file, I need to read the overlap matrix as well, right?

Yes, you need the overlap matrix as well.