@pfebrer 2021-11-09
> Yes, I can't see why not. Except that the bandstructure object requires a parent with cell information.

So just pass the list of points to the `point` argument and set `division` to 1? It would be nice to have something to detect corners for this case, wouldn't it? :smile: